chore: consolidate postProcess tests into test/types
