+++ test/types/int.spec.ts
