+++ test/types/payload-types.ts
@@ -1,4 +1,7 @@
 /* tslint:disable */
+
+export type SecondGeneric<K, V> = { key: K; value: V };
+export type TestPluginGeneric<T> = { value: T };
 /* eslint-disable */
 /**
  * This file was automatically generated by Payload.
@@ -116,6 +119,9 @@ export interface Config {
     settings: SettingsSelect<false> | SettingsSelect<true>;
   };
   locale: null;
+  widgets: {
+    collections: CollectionsWidget;
+  };
   strictDraftTypes: true;
   user: User;
   jobs: {
@@ -481,6 +487,16 @@ export interface SettingsSelect<T extends boolean = true> {
   updatedAt?: T;
   createdAt?: T;
   globalType?: T;
+}
+/**
+ * This interface was referenced by `Config`'s JSON-Schema
+ * via the `definition` "collections_widget".
+ */
+export interface CollectionsWidget {
+  data?: {
+    [k: string]: unknown;
+  };
+  width: 'full';
 }
 /**
  * This interface was referenced by `Config`'s JSON-Schema
